requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.2
django-cors-headers>=4.1.0
pillow==11.1.0
python-dotenv==1.1.0
whitenoise==6.9.0
drf-yasg==1.21.10
gunicorn==23.0.0
djangorestframework-simplejwt==5.5.0
whitenoise==6.9.0
psycopg2-binary==2.9.10
unidecode==1.4.0
requests==2.32.3
pytest
pytest-django